ShopmeWebParent/ShopmeFrontEnd/src/test/java/com/brooklyn/shopme/product/ProductRepositoryTest.java
package com.brooklyn.shopme.product;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;

import com.brooklyn.shopme.common.entity.Product;

@DataJpaTest
@AutoConfigureTestDatabase(replace = Replace.NONE)
public class ProductRepositoryTest {
	@Autowired
	private ProductRepository productRepository;
	@Test
	public void testGetProductByAlias(){
		String alias = "Acer-Chromebook-Spin-11-CP311-1H-C5PN";
		Product product = productRepository.findByAlias(alias);
		assertThat(product).isNotNull();
	}
}
